Fix warning about extraneous parens in an if statment

+++ src/benchmarks/mod.rs
@@ -25,7 +25,7 @@ use ruleset::KgsChinese;
 
 use time::get_time;
 
-pub fn pps(size: u8, runtime: uint) {
+pub fn pps(size: u8, runtime: i64) {
     let game = Game::new(size, 6.5, KgsChinese);
     let board = game.board();
     let playout_engine = Playout::new(board);
@@ -36,7 +36,7 @@ pub fn pps(size: u8, runtime: uint) {
         playout_engine.run();
         counter += 1;
 
-        if(get_time().sec - start >= runtime as i64) {
+        if (get_time().sec - start) >= runtime {
             break;
         }
     }
